rename organisms directory to molecules
