+++ apps/client/pages/created-exams.tsx
@@ -1,8 +1,8 @@
-import type { NextPage } from 'next';
-import Head from 'next/head';
 import { ExamTile } from 'components/atoms/ExamTile';
 import { SearchInput } from 'components/atoms/SearchInput';
-import { Table } from 'components/organisms/Table/Table';
+import { Table } from 'components/molecules/Table/Table';
+import type { NextPage } from 'next';
+import Head from 'next/head';
 import Link from 'next/link';
 
 const CreatedExams: NextPage = () => {
